+++ NEWS.md
@@ -1,6 +1,10 @@
-# distributions3 (development version)
+# distributions3 0.1.1.9000 (development version)
+To be released as distributions 0.2.0
 
+
+* Added a plotting generic for univariate distributions (@paulnorthrop, PR #56)
 * Added support for the Generalised Extreme Value (GEV), Frechet, Gumbel, reversed Weibull and Generalised Pareto (GP) distributions (@paulnorthrop, PR #52)
+
 
 # distributions 0.1.1